AO3-4359 - Split out API specs for better clarity
